Solved riddle for AoC 2024 day 22 part 2.

+++ aoc2024/aoc2024-kotlin/src/test/kotlin/de/havox_design/aoc2024/day22/MonkeyMarketTest.kt
@@ -15,7 +15,7 @@ class MonkeyMarketTest {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart2")
-    fun testProcessPart2(filename: String, expectedResult: Long) =
+    fun testProcessPart2(filename: String, expectedResult: Int) =
         MonkeyMarket(filename).processPart2().shouldBe(expectedResult)
 
     companion object {
@@ -28,7 +28,7 @@ class MonkeyMarketTest {
         @JvmStatic
         private fun getDataForTestProcessPart2(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2024/day22/day22part2sample.txt", 23L)
+                Arguments.of("de/havox_design/aoc2024/day22/day22part2sample.txt", 23)
             )
     }
 }
